added schema file for structure
